Refactoring the code around the tray area in order to improve usage
