+++ app/browser/pageTitleNotifications.js
